로그인 / 로그아웃 기능, 글 삭제 수정은 나만 가능, 아이디 수정 기능 추가
자기 소개 기능도 있음
아이디 삭제도 있음

+++ assignment_template/src/delete.php
@@ -1,10 +1,25 @@
 <?php
 $conn = mysqli_connect('db', 'root', 'rootpass', 'open');
+session_start();
+if (!isset($_SESSION['user_id'])) {
+    echo '로그인이 필요합니다. <a href="login.php">로그인</a>';
+    exit;
+}
+
 settype($_GET['id'], 'integer');
 
 $filtered = array(
     'id' => mysqli_real_escape_string($conn, $_GET['id'])
 );
+$sql = "SELECT author FROM topic WHERE id = {$filtered['id']}";
+$result = mysqli_query($conn, $sql);
+$row = mysqli_fetch_array($result);
+
+if ($row['author'] != $_SESSION['user_id'])
+ {
+    echo "본인 글만 삭제할 수 있습니다.";
+    exit;
+}
 
 $sql = "DELETE FROM topic WHERE id = {$filtered['id']}";
